+++ chopper/example/main.dart
@@ -2,19 +2,16 @@ import 'package:chopper/chopper.dart';
 import 'definition.dart';
 
 main() async {
-  final chopper = new ChopperClient(
-      baseUrl: "http://localhost:8000",
-      converter: const JsonConverter(),
-      apis: [
-        // the generated service
-        new MyService()
-      ],
-      /* ResponseInterceptorFunc | RequestInterceptorFunc | ResponseInterceptor | RequestInterceptor */
-      interceptors: [
-        new Headers(const {"Content-Type": "application/json"}),
-      ]);
+  final chopper = ChopperClient(
+    baseUrl: "http://localhost:8000",
+    apis: [
+      // the generated service
+      MyService()
+    ],
+    jsonApi: true,
+  );
 
-  final myService = chopper.service(MyService) as MyService;
+  final myService = chopper.service<MyService>();
 
   final response = await myService.getMapResource("1");
   print(response.body);