+++ chopper/lib/src/request.dart
@@ -1,3 +1,4 @@
+import 'dart:async';
 import 'dart:convert';
 
 import 'package:meta/meta.dart';
@@ -8,8 +9,12 @@ class Request {
   final String method;
   final String url;
   final dynamic body;
+  final List<PartValue> parts;
   final Map<String, dynamic> parameters;
   final Map<String, String> headers;
+  final bool formUrlEncoded;
+  final bool multipart;
+  final bool json;
 
   const Request(
     this.method,
@@ -17,6 +22,10 @@ class Request {
     this.body,
     this.parameters: const {},
     this.headers: const {},
+    this.multipart: false,
+    this.formUrlEncoded,
+    this.json,
+    this.parts: const [],
   });
 
   String _getMethod(String method) {
@@ -41,39 +50,104 @@ class Request {
     Map<String, dynamic> parameters,
     Map<String, String> headers,
     Encoding encoding,
+    List<PartValue> parts,
+    bool json,
+    bool multipart,
+    bool formUrlEncoded,
   }) =>
-      new Request(
+      Request(
         method ?? this.method,
         url ?? this.url,
         body: body ?? this.body,
         parameters: parameters ?? this.parameters,
         headers: headers ?? this.headers,
+        parts: parts ?? this.parts,
+        json: json ?? this.json,
+        formUrlEncoded: formUrlEncoded ?? this.formUrlEncoded,
+        multipart: multipart ?? this.multipart,
       );
 
   Uri _buildUri(String baseUrl) {
+    var uri;
     if (!baseUrl.endsWith('/') && !url.startsWith('/')) {
-      return Uri.parse("$baseUrl/$url");
+      uri = Uri.parse("$baseUrl/$url");
+    } else {
+      uri = Uri.parse("$baseUrl$url");
     }
-    return Uri.parse("$baseUrl$url");
+
+    if (parameters.isNotEmpty) {
+      return uri.replace(
+        queryParameters: parameters.map((k, v) => MapEntry(k, "$v")),
+      );
+    }
+    return uri;
   }
 
-  http.BaseRequest toHttpRequest(String baseUrl) {
-    final uri = _buildUri(baseUrl).replace(
-      queryParameters: parameters.map((k, v) => new MapEntry(k, "$v")),
-    );
-    final baseRequest = new http.Request(_getMethod(method), uri);
+  Future<http.BaseRequest> _toMultipartRequest(
+    String method,
+    Uri uri, {
+    bool formUrlEncodedApi: false,
+  }) async {
+    final baseRequest = http.MultipartRequest(method, uri);
     baseRequest.headers.addAll(headers);
-    if (body != null) {
-      if (body is String) {
-        baseRequest.body = body as String;
-      } else if (body is List) {
-        baseRequest.bodyBytes = (body as List<int>).cast<int>();
-      } else if (body is Map) {
-        baseRequest.bodyFields = (body as Map).cast<String, String>();
+
+    if (formUrlEncodedApi || formUrlEncoded == true) {
+      baseRequest.headers["Content-Type"] = 'application/x-www-form-urlencoded';
+    }
+
+    for (final part in parts) {
+      if (part is PartFile) {
+        if (part.value is List<int>) {
+          baseRequest.files.add(
+            http.MultipartFile.fromBytes(part.name, part.value),
+          );
+        } else if (part.value is String) {
+          baseRequest.files.add(
+            await http.MultipartFile.fromPath(part.name, part.value),
+          );
+        }
       } else {
-        throw new ArgumentError('Invalid request body "${body}".');
+        baseRequest.fields[part.name] = part.value?.toString();
       }
     }
+    return baseRequest;
+  }
+
+  http.BaseRequest toHttpRequest(
+    String baseUrl, {
+    bool formUrlEncodedApi: false,
+  }) {
+    final uri = _buildUri(baseUrl);
+
+    var baseRequest;
+
+    if (multipart) {
+      baseRequest = _toMultipartRequest(
+        _getMethod(method),
+        uri,
+        formUrlEncodedApi: formUrlEncodedApi,
+      );
+    } else {
+      baseRequest = http.Request(_getMethod(method), uri);
+      baseRequest.headers.addAll(headers);
+
+      if (formUrlEncodedApi || formUrlEncoded == true) {
+        baseRequest.headers["Content-Type"] =
+            'application/x-www-form-urlencoded';
+      }
+      if (body != null) {
+        if (body is String) {
+          baseRequest.body = body as String;
+        } else if (body is List) {
+          baseRequest.bodyBytes = (body as List<int>).cast<int>();
+        } else if (body is Map) {
+          baseRequest.bodyFields = (body as Map).cast<String, String>();
+        } else {
+          throw ArgumentError('Invalid request body "${body}".');
+        }
+      }
+    }
+
     return baseRequest;
   }
 }
@@ -84,4 +158,20 @@ class HttpMethod {
   static const String Put = "PUT";
   static const String Delete = "DELETE";
   static const String Patch = "PATCH";
+}
+
+@immutable
+class PartValue<T> {
+  final T value;
+  final String name;
+
+  const PartValue(
+    this.name,
+    this.value,
+  );
+}
+
+@immutable
+class PartFile<T> extends PartValue<T> {
+  PartFile(String name, T value) : super(name, value);
 }