+++ example/bin/definition/definition.chopper.dart
@@ -17,20 +17,20 @@ class MyService extends ChopperService implements MyServiceDefinition {
     final url = '/';
     final params = {'id': id};
     final request = new Request('GET', url, parameters: params);
-    return client.send<Map>(request, responseType: Map);
+    return client.send<Map>(request);
   }
 
   Future<Response<Resource>> getTypedResource() {
     final url = '/';
     final headers = {'foo': 'bar'};
     final request = new Request('GET', url, headers: headers);
-    return client.send<Resource>(request, responseType: Resource);
+    return client.send<Resource>(request);
   }
 
   Future<Response<Resource>> newResource(Resource resource, {String name}) {
     final url = '/';
     final headers = {'name': name};
     final request = new Request('POST', url, body: resource, headers: headers);
-    return client.send<Resource>(request, responseType: Resource);
+    return client.send<Resource>(request);
   }
 }